roughly working version of normalization stylesheet.
this does two things:
1. creates a reference sequence of all elements with ids and the contents of
   those elements extends down only to the next element with an id or idref.
   within these templates, any nested elements with an id are changed to be
   references using an idref attribute

2. create a reference copy of the source document that replaces all
   occurrences of elements with ids with a reference using an idref
   attribute.

(the actual stylesheet logic is a little more complicated because it has
to distinguish between "top level" elements and nested elements.  the
former are main element in the reference and need to keep their ids. the
nested elements should have their ids replaced with idrefs and child nodes
are lopped off.)